// src/types/Enums/SecondarySkills.ts
export enum SecondarySkills {
  ARCHERY = 'Меткость',
  LEADERSHIP = 'Лидерство',
  DEFENSE = 'Защита',
  NAVIGATION = 'Навигация',
  ESTATES = 'Экономика',
  OFFENSE = 'Нападение',
  BALLISTA = 'Артиллерия',
  TACTICS = 'Тактика',
  WISDOM = 'Мудрость',
  FIRST_AID = 'Первая Помощь',
  DIPLOMACY = 'Дипломатия',
  INTELLIGENCE = 'Интеллект',
  MYSTICISM = 'Мистицизм',
  LEARNING = 'Учение',
  EAGLE_EYE = 'Зоркость',
  LOGISTICS = 'Логистика',
  AIR_MAGIC = 'Магия Воздуха',
  WATER_MAGIC = 'Магия Воды',
  EARTH_MAGIC = 'Магия Земли',
  FIRE_MAGIC = 'Магия Огня',
  RESISTANCE = 'Сопротивление',
  SCOUTING = 'Разведка',
  SCOLAR = 'Ученый',
  SORCERY = 'Колдовство',
  PATH_FINDING= 'Поиск Пути',
  BALISTICS = 'Катапульта',
  NECROMANCY = 'Некромантия',
  LUCK = 'Удача',
}